Hide "Manage" menu in nav-bar if the role is not a Director.
Reorganize menu item in nav-bar. Passport authenticate return user_id and role_id. Add parse in client controller to obtain the user_id and role_id. Perform check and hide "Manage" menu in nav-bar if the role is not a Director so that member would not be able to access this functionality.

Change href to ui-sref link reference in user_event director.